wip testing single GPU training

+++ training_utils.py
@@ -9,13 +9,14 @@ def get_batch(dataset: np.array, batch_size: int, context_length: int, device: s
 
     sample_ids = torch.randint(low=0, high=set_size-context_length, size=(batch_size,))
 
-    inputs = torch.empty(batch_size, context_length)
-    targets = torch.empty(batch_size, context_length)
+    inputs = torch.empty(batch_size, context_length, dtype=torch.long)
+    targets = torch.empty(batch_size, context_length, dtype=torch.long)
 
     for batch_id, sample_id in zip(range(batch_size), sample_ids):
-        inputs[batch_id] = torch.tensor(dataset[sample_id:sample_id+context_length])
+        inputs[batch_id] = torch.tensor(
+            dataset[sample_id:sample_id+context_length], dtype=torch.long)
         targets[batch_id] = torch.tensor(
-            dataset[sample_id+1:sample_id+1+context_length])
+            dataset[sample_id+1:sample_id+1+context_length], dtype=torch.long)
 
     return (inputs.to(device), targets.to(device))
